Fixed #3, #4

+++ app/Http/Controllers/CategoryController.php
@@ -140,13 +140,23 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $noSubcategories = $category->subcategories()->count();
+        if ($noSubcategories > 0 ) {
+            return response([
+                "status" => "failed",
+                "flash" => [
+                    "message" => "Can not delete as this category has " . $noSubcategories . " sub-categories under it.",
+                    "type" => "warning"
+                ],
+            ], 422);
+        }
+
         $category->delete();
 
         return [
             "status" => "success",
             "flash" => ["message" => "Category \"" . $category->name . "\" deleted"],
             "category_id" => $category->id
-          
         ];
     }
 }
